RDM-13146 Return CategoriesAndDocuments after document category update

+++ src/main/java/uk/gov/hmcts/ccd/domain/model/casefileview/CategoriesAndDocuments.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Value
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
 public class CategoriesAndDocuments {
-    Long caseVersion;
+    Integer caseVersion;
     List<Category> categories;
     List<Document> uncategorisedDocuments;
 }
